+++ packages/injected/package.json
@@ -18,6 +18,7 @@
   "license": "MIT",
   "devDependencies": {
     "@ava/typescript": "^2.0.0",
+    "@ethersproject/providers": "^5.5.0",
     "@swc/cli": "^0.1.51",
     "@swc/core": "^1.2.92",
     "@types/lodash.uniqby": "^4.7.6",
