+++ packages/injected/src/index.ts
@@ -1,18 +1,16 @@
 import uniqBy from 'lodash.uniqby'
 
-import type {
-  WalletInit,
-  InjectedWalletOptions,
-  CustomWindow
-} from '@web3-onboard/common'
-
-import { ProviderLabel } from '@web3-onboard/common'
+import type { WalletInit } from '@web3-onboard/common'
+import type { InjectedWalletOptions, CustomWindow } from './types'
+import { ProviderLabel } from './types'
 
 import standardWallets from './wallets.js'
 import { remove } from './helpers.js'
 import { validateWalletOptions } from './validation.js'
 
 declare const window: CustomWindow
+
+export { ProviderIdentityFlag, ProviderLabel } from './types'
 
 function injected(options?: InjectedWalletOptions): WalletInit {
   if (typeof window === 'undefined') return () => null
